[device] Rename common.py to exception.py
