+++ tools/device/modm/__init__.py
@@ -5,16 +5,16 @@ Device Platform Generator
 """
 
 from . import pkg
-from . import common
+from . import exception
 from . import device_file
 from . import device_identifier
 from . import device
 from . import parser
 
 from .pkg import naturalkey
-from .common import ParserException
+from .exception import ParserException
 
-__all__ = ['common', 'device_file', 'device_identifier', 'device', 'parser', 'pkg']
+__all__ = ['exception', 'device_file', 'device_identifier', 'device', 'parser', 'pkg']
 
 __author__ = "Fabian Greif"
 __copyright__ = "Fabian Greif"
